+++ systemUptime@KopfDesDaemons/files/systemUptime@KopfDesDaemons/CHANGELOG.md
@@ -12,6 +12,8 @@
 ## [1.1.0] - 06.02.2026
 
 - use local date and time format
+- use the "Use 24H clock" cinnamon setting
 - use St.Icon for the clock icon
 - use file_get_contents() to get starttime and uptime
+- update screenshot
 - add more comments